Select default file path based on parsing mode
Prior to this commit the default file path was always '290.rec',
regardless of whether you were trying to parse a Cave Story+ '.dat' file
or a freeware Cave Story '.rec' file. This commit changes this behavior,
defaulting to 'Profile.dat' when the Cave Story+ flag is set, and
otherwise defaulting to '290.rec'.