Attempt to catch misuse of of 'csplus' flag
This commit adds a check to see whether the use may either be enabling
the 'csplus' flag when trying to parse a freeware Cave Story '.rec'
file, or not enabling it when trying to parse a Cave Story+ '.dat' file.

In either of these cases a helpful message will be displayed telling the
user how to set or un-set the flag.